refactor: move `clear_pending_commit` from `CentralContext` to `ConversationGuard`
