+++ crypto/src/mls/conversation/conversation_guard/merge.rs
@@ -1,14 +1,15 @@
-//! A MLS group can be merged (aka committed) when it has a pending commit. The latter is a commit
-//! we created which is still waiting to be "committed". By doing so, we will apply all the
-//! modifications present in the commit to the ratchet tree and also persist the new group in the
-//! keystore. Like this, even if the application crashes we will be able to restore.
-//!
-//! This table summarizes when a MLS group can be merged:
-//!
-//! | can be merged ?   | 0 pend. Commit | 1 pend. Commit |
-//! |-------------------|----------------|----------------|
-//! | 0 pend. Proposal  | ❌              | ✅              |
-//! | 1+ pend. Proposal | ❌              | ✅              |
+//! A MLS group is a distributed object scattered across many parties. We use a Delivery Service
+//! to orchestrate those parties. So when we create a commit, a mutable operation, it has to be
+//! validated by the Delivery Service. But it might occur that another group member did the
+//! exact same thing at the same time. So if we arrive second in this race, we must "rollback" the commit
+//! we created and accept ("merge") the other one.
+//! A client would
+//! * Create a commit
+//! * Send the commit to the Delivery Service
+//! * When Delivery Service responds
+//!     * 200 OK --> use [CentralContext::commit_accepted] to merge the commit
+//!     * 409 CONFLICT --> do nothing. [CentralContext::decrypt_message] will restore the proposals not committed
+//!     * 5xx --> retry
 
 use openmls::prelude::MlsGroupStateError;
 
@@ -47,6 +48,31 @@ impl ConversationGuard {
             })?;
         conversation.persist_group_when_changed(&keystore, true).await?;
         Ok(())
+    }
+
+    /// Allows to remove a pending commit. Use this when backend rejects the commit
+    /// you just sent e.g. if permissions have changed meanwhile.
+    ///
+    /// **CAUTION**: only use this when you had an explicit response from the Delivery Service
+    /// e.g. 403. Do not use otherwise e.g. 5xx responses, timeout etc..
+    /// **DO NOT** use when Delivery Service responds 409, pending state will be renewed
+    /// in [CentralContext::decrypt_message]
+    ///
+    /// # Arguments
+    /// * `conversation_id` - the group/conversation id
+    ///
+    /// # Errors
+    /// When the conversation is not found or there is no pending commit
+    pub(crate) async fn clear_pending_commit(&mut self) -> Result<()> {
+        let keystore = self.mls_provider().await?.keystore();
+        let mut conversation = self.conversation_mut().await;
+        if conversation.group.pending_commit().is_some() {
+            conversation.group.clear_pending_commit();
+            conversation.persist_group_when_changed(&keystore, true).await?;
+            Ok(())
+        } else {
+            Err(Error::PendingCommitNotFound)
+        }
     }
 }
 
@@ -219,6 +245,100 @@ mod tests {
                     // for the Update proposal
                     let after_clear_proposal = cc.context.count_entities().await;
                     assert_eq!(init, after_clear_proposal);
+                })
+            })
+            .await
+        }
+    }
+
+    mod clear_pending_commit {
+        use super::*;
+
+        #[apply(all_cred_cipher)]
+        #[wasm_bindgen_test]
+        pub async fn should_remove_commit(case: TestCase) {
+            run_test_with_client_ids(case.clone(), ["alice"], move |[alice_central]| {
+                Box::pin(async move {
+                    let id = conversation_id();
+                    alice_central
+                        .context
+                        .new_conversation(&id, case.credential_type, case.cfg.clone())
+                        .await
+                        .unwrap();
+                    assert!(alice_central.pending_commit(&id).await.is_none());
+
+                    alice_central.create_unmerged_commit(&id).await;
+                    assert!(alice_central.pending_commit(&id).await.is_some());
+                    alice_central
+                        .context
+                        .conversation_guard(&id)
+                        .await
+                        .unwrap()
+                        .clear_pending_commit()
+                        .await
+                        .unwrap();
+                    assert!(alice_central.pending_commit(&id).await.is_none());
+                })
+            })
+            .await
+        }
+
+        #[apply(all_cred_cipher)]
+        #[wasm_bindgen_test]
+        pub async fn should_fail_when_pending_commit_absent(case: TestCase) {
+            run_test_with_client_ids(case.clone(), ["alice"], move |[alice_central]| {
+                Box::pin(async move {
+                    let id = conversation_id();
+                    alice_central
+                        .context
+                        .new_conversation(&id, case.credential_type, case.cfg.clone())
+                        .await
+                        .unwrap();
+                    assert!(alice_central.pending_commit(&id).await.is_none());
+                    let clear = alice_central
+                        .context
+                        .conversation_guard(&id)
+                        .await
+                        .unwrap()
+                        .clear_pending_commit()
+                        .await;
+                    assert!(matches!(clear.unwrap_err(), Error::PendingCommitNotFound))
+                })
+            })
+            .await
+        }
+
+        #[apply(all_cred_cipher)]
+        #[wasm_bindgen_test]
+        pub async fn should_clean_associated_key_material(case: TestCase) {
+            run_test_with_client_ids(case.clone(), ["alice"], move |[cc]| {
+                Box::pin(async move {
+                    let id = conversation_id();
+                    cc.context
+                        .new_conversation(&id, case.credential_type, case.cfg.clone())
+                        .await
+                        .unwrap();
+                    assert!(cc.pending_commit(&id).await.is_none());
+
+                    let init = cc.context.count_entities().await;
+
+                    cc.create_unmerged_commit(&id).await;
+                    assert!(cc.pending_commit(&id).await.is_some());
+
+                    cc.context
+                        .conversation_guard(&id)
+                        .await
+                        .unwrap()
+                        .clear_pending_commit()
+                        .await
+                        .unwrap();
+                    assert!(cc.pending_commit(&id).await.is_none());
+
+                    // This whole flow should be idempotent.
+                    // Here we verify that we are indeed deleting the `EncryptionKeyPair` created
+                    // for the Update commit
+                    let after_clear_commit = cc.context.count_entities().await;
+                    assert_eq!(init, after_clear_commit);
                 })
             })
             .await
