+++ crypto/src/mls/conversation/merge.rs
@@ -16,11 +16,8 @@ use openmls_traits::OpenMlsCryptoProvider;
 
 use mls_crypto_provider::MlsCryptoProvider;
 
-use super::{Error, Result};
-use crate::{MlsError, context::CentralContext, mls::MlsConversation};
-
-#[cfg(test)]
-use crate::{RecursiveError, mls::ConversationId};
+use super::Result;
+use crate::{MlsError, mls::MlsConversation};
 
 /// Abstraction over a MLS group capable of merging a commit
 impl MlsConversation {
@@ -44,61 +41,6 @@ impl MlsConversation {
 
         Ok(())
     }
-
-    /// see [CentralContext::clear_pending_commit]
-    #[cfg_attr(test, crate::durable)]
-    pub(crate) async fn clear_pending_commit(&mut self, backend: &MlsCryptoProvider) -> Result<()> {
-        if self.group.pending_commit().is_some() {
-            self.group.clear_pending_commit();
-            self.persist_group_when_changed(&backend.keystore(), true).await?;
-            Ok(())
-        } else {
-            Err(Error::PendingCommitNotFound)
-        }
-    }
-}
-
-/// A MLS group is a distributed object scattered across many parties. We use a Delivery Service
-/// to orchestrate those parties. So when we create a commit, a mutable operation, it has to be
-/// validated by the Delivery Service. But it might occur that another group member did the
-/// exact same thing at the same time. So if we arrive second in this race, we must "rollback" the commit
-/// we created and accept ("merge") the other one.
-/// A client would
-/// * Create a commit
-/// * Send the commit to the Delivery Service
-/// * When Delivery Service responds
-///     * 200 OK --> use [CentralContext::commit_accepted] to merge the commit
-///     * 409 CONFLICT --> do nothing. [CentralContext::decrypt_message] will restore the proposals not committed
-///     * 5xx --> retry
-impl CentralContext {
-    /// Allows to remove a pending commit. Use this when backend rejects the commit
-    /// you just sent e.g. if permissions have changed meanwhile.
-    ///
-    /// **CAUTION**: only use this when you had an explicit response from the Delivery Service
-    /// e.g. 403. Do not use otherwise e.g. 5xx responses, timeout etc..
-    /// **DO NOT** use when Delivery Service responds 409, pending state will be renewed
-    /// in [CentralContext::decrypt_message]
-    ///
-    /// # Arguments
-    /// * `conversation_id` - the group/conversation id
-    ///
-    /// # Errors
-    /// When the conversation is not found or there is no pending commit
-    #[cfg(test)]
-    #[cfg_attr(test, crate::idempotent)]
-    pub(crate) async fn clear_pending_commit(&self, conversation_id: &ConversationId) -> Result<()> {
-        self.get_conversation(conversation_id)
-            .await?
-            .write()
-            .await
-            .clear_pending_commit(
-                &self
-                    .mls_provider()
-                    .await
-                    .map_err(RecursiveError::root("getting mls provider"))?,
-            )
-            .await
-    }
 }
 
 #[cfg(test)]
@@ -106,8 +48,6 @@ mod tests {
     use wasm_bindgen_test::*;
 
     use crate::test_utils::*;
-
-    use super::*;
 
     wasm_bindgen_test_configure!(run_in_browser);
 
@@ -201,96 +141,6 @@ mod tests {
 
                     let final_count = alice_central.context.count_entities().await;
                     assert_eq!(initial_count, final_count);
-                })
-            })
-            .await
-        }
-    }
-
-    mod clear_pending_commit {
-        use super::*;
-
-        #[apply(all_cred_cipher)]
-        #[wasm_bindgen_test]
-        pub async fn should_remove_commit(case: TestCase) {
-            run_test_with_client_ids(case.clone(), ["alice"], move |[alice_central]| {
-                Box::pin(async move {
-                    let id = conversation_id();
-                    alice_central
-                        .context
-                        .new_conversation(&id, case.credential_type, case.cfg.clone())
-                        .await
-                        .unwrap();
-                    assert!(alice_central.pending_commit(&id).await.is_none());
-
-                    alice_central.create_unmerged_commit(&id).await;
-                    assert!(alice_central.pending_commit(&id).await.is_some());
-                    alice_central.context.clear_pending_commit(&id).await.unwrap();
-                    assert!(alice_central.pending_commit(&id).await.is_none());
-                })
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        #[wasm_bindgen_test]
-        pub async fn should_fail_when_conversation_not_found(case: TestCase) {
-            use crate::LeafError;
-
-            run_test_with_client_ids(case.clone(), ["alice"], move |[alice_central]| {
-                Box::pin(async move {
-                    let id = conversation_id();
-                    let clear = alice_central.context.clear_pending_commit(&id).await;
-                    assert!(matches!(clear.unwrap_err(), Error::Leaf(LeafError::ConversationNotFound(conv_id)) if conv_id == id))
-                })
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        #[wasm_bindgen_test]
-        pub async fn should_fail_when_pending_commit_absent(case: TestCase) {
-            run_test_with_client_ids(case.clone(), ["alice"], move |[alice_central]| {
-                Box::pin(async move {
-                    let id = conversation_id();
-                    alice_central
-                        .context
-                        .new_conversation(&id, case.credential_type, case.cfg.clone())
-                        .await
-                        .unwrap();
-                    assert!(alice_central.pending_commit(&id).await.is_none());
-                    let clear = alice_central.context.clear_pending_commit(&id).await;
-                    assert!(matches!(clear.unwrap_err(), Error::PendingCommitNotFound))
-                })
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        #[wasm_bindgen_test]
-        pub async fn should_clean_associated_key_material(case: TestCase) {
-            run_test_with_client_ids(case.clone(), ["alice"], move |[cc]| {
-                Box::pin(async move {
-                    let id = conversation_id();
-                    cc.context
-                        .new_conversation(&id, case.credential_type, case.cfg.clone())
-                        .await
-                        .unwrap();
-                    assert!(cc.pending_commit(&id).await.is_none());
-
-                    let init = cc.context.count_entities().await;
-
-                    cc.create_unmerged_commit(&id).await;
-                    assert!(cc.pending_commit(&id).await.is_some());
-
-                    cc.context.clear_pending_commit(&id).await.unwrap();
-                    assert!(cc.pending_commit(&id).await.is_none());
-
-                    // This whole flow should be idempotent.
-                    // Here we verify that we are indeed deleting the `EncryptionKeyPair` created
-                    // for the Update commit
-                    let after_clear_commit = cc.context.count_entities().await;
-                    assert_eq!(init, after_clear_commit);
                 })
             })
             .await
